pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
[project]
name = "closer"
version = "0.0.1"
authors = [
  { name="Cloud956", email="placeholder@email.com" },
]
description = "NLP pipeline for most legal tasks"
readme = "README.md"
license = {file = "LICENSE.txt"}

dependencies = [
    "absl-py==1.4.0",
    "aiohttp==3.8.4",
    "aiosignal==1.3.1",
    "async-timeout==4.0.2",
    "attrs==23.1.0",
    "beautifulsoup4==4.12.2",
    "blis==0.7.9",
    "boto3==1.26.135",
    "botocore==1.29.135",
    "bs4==0.0.1",
    "cachetools==5.3.0",
    "catalogue==2.0.8",
    "certifi==2023.5.7",
    "charset-normalizer==3.1.0",
    "click==8.1.3",
    "confection==0.0.4",
    "cymem==2.0.7",
    "Cython==0.29.34",
    "dataclasses-json==0.5.7",
    "docopt==0.6.2",
    "et-xmlfile==1.1.0",
    "filelock==3.12.0",
    "frozenlist==1.3.3",
    "fsspec==2023.5.0",
    "future==0.18.3",
    "google-auth==2.18.0",
    "google-auth-oauthlib==1.0.0",
    "greenlet==2.0.2",
    "grpcio==1.54.2",
    "huggingface-hub==0.14.1",
    "idna==3.4",
    "inflect==6.0.4",
    "Jinja2==3.1.2",
    "jmespath==1.0.1",
    "joblib==1.2.0",
    "langchainplus-sdk==0.0.7",
    "langcodes==3.3.0",
    "llama-index==0.8.42",
    "lmdb==1.4.1",
    "Markdown==3.4.3",
    "MarkupSafe==2.1.2",
    "marshmallow==3.19.0",
    "marshmallow-enum==1.5.1",
    "mpmath==1.3.0",
    "multidict==6.0.4",
    "murmurhash==1.0.9",
    "mypy-extensions==1.0.0",
    "networkx==3.1",
    "nltk==3.8.1",
    "num2words==0.5.12",
    "numexpr==2.8.4",
    "numpy==1.24.3",
    "oauthlib==3.2.2",
    "openai==0.27.8",
    "openapi-schema-pydantic==1.2.4",
    "openpyxl==3.1.2",
    "packaging==23.1",
    "pandas==2.0.1",
    "pathy==0.10.1",
    "preshed==3.0.8",
    "prettyprint==0.1.5",
    "protobuf==4.23.0",
    "pyasn1==0.5.0",
    "pyasn1-modules==0.3.0",
    "pydantic==1.10.7",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "pytz==2023.3",
    "PyYAML==6.0",
    "quantulum3==0.9.0",
    "regex==2023.5.5",
    "requests==2.29.0",
    "requests-oauthlib==1.3.1",
    "rsa==4.9",
    "s3transfer==0.6.1",
    "six==1.16.0",
    "smart-open==6.3.0",
    "soupsieve==2.4.1",
    "spacy==3.5.3",
    "spacy-entity-linker==1.0.3",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.4",
    "SQLAlchemy==2.0.15",
    "srsly==2.4.6",
    "sympy==1.12",
    "tagme==0.1.3",
    "tenacity==8.2.2",
    "tensorboard==2.13.0",
    "tensorboard-data-server==0.7.0",
    "thinc==8.1.10",
    "tiktoken==0.4.0",
    "tokenizers==0.13.3",
    "torch==2.0.1",
    "tqdm==4.65.0",
    "transformers==4.29.2",
    "typer==0.7.0",
    "typing-inspect==0.8.0",
    "typing_extensions==4.5.0",
    "tzdata==2023.3",
    "ujson==5.7.0",
    "Unidecode==1.3.6",
    "urllib3==1.26.15",
    "wasabi==1.1.1",
    "Werkzeug==2.3.4",
    "XlsxWriter==3.1.2",
    "yarl==1.9.2",
    "stanford_openie==1.3.1",
    "legal_openai"
]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.scripts]
build = "src.__init__:main"

[project.urls]